+++ src/app/gui/show-code-dialog/tabs/angular-tab.component.ts
@@ -2,6 +2,7 @@ import {Component, Input} from '@angular/core'
 import {ComponentParams} from "../../../model/component-params";
 import {copyToClipboard} from "../../../shared/utils";
 import packageInfo  from '../../../../../package.json'
+import {ToastService} from "../../../shared/services/toast/toast.service";
 
 @Component({
   selector: 'app-angular-tab',
@@ -17,7 +18,14 @@ void defineCustomElements(window)`
   public get link3(): string {
     return `<script type='module' src='https://unpkg.com/browse/animated-banner@${packageInfo.version}/dist/animations/${this.data.animation}.js'></script>`
   }
+
+  constructor(
+    public toast: ToastService
+  ) {
+  }
+
   public copy(text): void {
     copyToClipboard(text)
+    this.toast.info('Copied to clipboard!')
   }
 }